paint-timing/fcp-only/fcp-video-frame.html

<!DOCTYPE html>
<head>
<title>Performance Paint Timing Test: FCP due to loaded video frame</title>
<style>#log {display: none}</style>
</head>
<body>
<script src="../resources/utils.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<video id="video" autoplay></video>
</body>

<footer>
<script>
  promise_test(async t => {
    try {
        assert_precondition(window.PerformancePaintTiming, "Paint Timing isn't supported.");
        await new Promise(r => window.addEventListener('load', r))
        await assertNoFirstContentfulPaint(t)
        const video = document.getElementById('video')
        video.setAttribute('src', '/media/test.mp4')
        await new Promise(resolve => {
          video.oncanplay = resolve
        })
        await assertFirstContentfulPaint(t)
    } catch (e) {
      document.getElementById('log').style.display = 'block'
      throw e
    } finally {
      document.getElementById('log').style.display = 'block'
    }
  }, 'Video should become contentful when first frame is loaded')
</script>
<footer>
</html>
